backend/general/splink/requirements.txt
splink>=4.0
duckdb
cleanco
psycopg2-binary
pyarrow
python-dotenv
pandas
